Add ability to codegen for individual functions
Currently useful for tests, but later may be used for incremental compilation and/or hot reloading
